+++ CHANGELOG.md
@@ -5,6 +5,10 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 
+### Fixed
+
+- Ignore not running pods when maintaining runner pods. (#143)
+
 ## [0.6.2] - 2022-03-23
 
 ### Changed